Added missing boxart
Added the box art for ever Professor Layton game that didn't have one yet. I added the normal box art for the demo and kiosk versions as well. It seems those versions are usually distributed in small plastic boxes without any proper packaging, and have a 'NOT FOR RESALE' label printed on the cartridge.