Reverted Etrian Odyssey to its first cover
Sorry, I wasn't sure if I should replace the first one with the slightly blurry one. Next time if a change leads to quality loss I will make a separate PR so it can be denied by itself. If you want the initial cover here it is, if not just delete this PR. (The only difference I see is, that the new cover might be a bit brighter.)